refactor: move project quality to projects dashboard
